fix: Respect maxVCpus in miniwdl contexts
We currently don't set the maxVCpus from the contextParameters when
creating MiniWDL Batch resources. This commit ensures that we do.
This is possible because the code instantiating these resources is
duplicated. A refactor to remove this duplication will be done in a
subsequent commit, so as not to pollute the patch release.