Remove Chdir when looking for project files
We currently search for the agc-project.yaml file and chdir into the
project directory in case AGC is run from a folder below the top-level
project folder. This means that when looking for files in the args command,
the CLI is not looking in the users current directory, meaning they must
specify --args files relative to the project.yaml rather than their
working directory. This commit changes AGC to prepend the project directory
to any workflow lookups, so the working directory remains the same as the users.