Test the exact validation sequence implemented for adapter logs
If we always make the context manager believable enough to actually return
contexts, we fail the tests because the check for whatever kind of error we're
testing happens to come before the implementation accesses the context manager,
and the mock gets upset because the method we told it to expect is never
called.

There might be a way to allow but not require a method on a mock to be called,
but I don't know it. So instead we just support the actual call sequence
observed, coupling the business logic and the test cases, and if the order of
the checks in the business logic ever changes we will have to change the test
cases to accomodate it.